forbid compilation for target_pointer_width < 32
This is done so that conversions from u32 to usize are guaranteed to be correct.

This is technically not a breaking change: the only <32-bit target listed in the Platform Support page of the rustc book (https://doc.rust-lang.org/1.73.0/rustc/platform-support.html) - `msp430-none-elf` - does not support the standard library, so this library would never have compiled for it to begin with.